Extract Loader to separate module
